pyproject.toml
[project]
name = "autoagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.111.0",
    "google-genai>=0.3.0",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.13.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=2.7.0",
    "sqlmodel>=0.0.38",
    "streamlit>=1.35.0",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.44.0",
]
